+++ src/components/Events/EventComponent.vue
@@ -95,8 +95,8 @@
 
 <script>
   import EventItem from "./EventItem.vue";
-  import { useEventsStore } from "../stores/EventsStore.js";
-  import { useUserStore } from "../stores/UserStore.js";
+  import { useEventsStore } from "../../stores/EventsStore.js";
+  import { useUserStore } from "../../stores/UserStore.js";
   import { mapStores } from "pinia";
   export default {
     name: "EventComponent",
