+++ src/components/Events/EventSignupItem.vue
@@ -58,10 +58,10 @@
 
 <script>
   import EventItem from "./EventItem.vue";
-  import { useEventsStore } from "../stores/EventsStore.js";
-  import { useUserStore } from "../stores/UserStore.js";
+  import { useEventsStore } from "../../stores/EventsStore.js";
+  import { useUserStore } from "../../stores/UserStore.js";
   import { mapStores } from "pinia";
-  import EventSignUpDataService from "../services/eventsignup.js";
+  import EventSignUpDataService from "../../services/eventsignup.js";
   export default {
     name: "EventSignupItem",
     components: {
